+++ plugin-dev/Source/Sentry/Private/Apple/AppleSentryBreadcrumb.cpp
@@ -11,10 +11,10 @@
 
 FAppleSentryBreadcrumb::FAppleSentryBreadcrumb()
 {
-	BreadcrumbApple = [[SENTRY_APPLE_CLASS(SentryBreadcrumb) alloc] init];
+	BreadcrumbApple = [[SENTRY_APPLE_CLASS(SentryObjCBreadcrumb) alloc] init];
 }
 
-FAppleSentryBreadcrumb::FAppleSentryBreadcrumb(SentryBreadcrumb* breadcrumb)
+FAppleSentryBreadcrumb::FAppleSentryBreadcrumb(SentryObjCBreadcrumb* breadcrumb)
 {
 	BreadcrumbApple = breadcrumb;
 }
@@ -24,7 +24,7 @@ FAppleSentryBreadcrumb::~FAppleSentryBreadcrumb()
 	// Put custom destructor logic here if needed
 }
 
-SentryBreadcrumb* FAppleSentryBreadcrumb::GetNativeObject()
+SentryObjCBreadcrumb* FAppleSentryBreadcrumb::GetNativeObject()
 {
 	return BreadcrumbApple;
 }