+++ plugin-dev/Source/Sentry/Private/Apple/AppleSentryLog.h
@@ -6,17 +6,17 @@
 
 #include "Interface/SentryLogInterface.h"
 
-@class SentryLog;
+@class SentryObjCLog;
 
 class FAppleSentryLog : public ISentryLog
 {
 public:
 	FAppleSentryLog();
-	FAppleSentryLog(SentryLog* log);
+	FAppleSentryLog(SentryObjCLog* log);
 	FAppleSentryLog(const FString& body, ESentryLevel level);
 	virtual ~FAppleSentryLog() override;
 
-	SentryLog* GetNativeObject();
+	SentryObjCLog* GetNativeObject();
 
 	virtual void SetBody(const FString& body) override;
 	virtual FString GetBody() const override;
@@ -30,7 +30,7 @@ public:
 	virtual void AddAttributes(const TMap<FString, FSentryVariant>& attributes) override;
 
 private:
-	SentryLog* LogApple;
+	SentryObjCLog* LogApple;
 };
 
 typedef FAppleSentryLog FPlatformSentryLog;