+++ plugin-dev/Source/Sentry/Private/Apple/AppleSentrySubsystem.h
@@ -76,6 +76,8 @@ protected:
 	bool isGameLogAttachmentEnabled = false;
 	bool isSessionReplayAttachmentEnabled = false;
 
+	int32 maxAttachmentSize = 0;
+
 	FString PrevSessionReplayPath;
 };