+++ plugin-dev/Source/Sentry/Private/Apple/AppleSentryUser.cpp
@@ -11,10 +11,10 @@
 
 FAppleSentryUser::FAppleSentryUser()
 {
-	UserApple = [[SENTRY_APPLE_CLASS(SentryUser) alloc] init];
+	UserApple = [[SENTRY_APPLE_CLASS(SentryObjCUser) alloc] init];
 }
 
-FAppleSentryUser::FAppleSentryUser(SentryUser* user)
+FAppleSentryUser::FAppleSentryUser(SentryObjCUser* user)
 {
 	UserApple = user;
 }
@@ -24,7 +24,7 @@ FAppleSentryUser::~FAppleSentryUser()
 	// Put custom destructor logic here if needed
 }
 
-SentryUser* FAppleSentryUser::GetNativeObject()
+SentryObjCUser* FAppleSentryUser::GetNativeObject()
 {
 	return UserApple;
 }